Refactor image revision with background queue
